requirements.txt
absl-py==1.0.0
altair==4.2.0
appdirs==1.4.4
apturl==0.5.2
argon2-cffi==20.1.0
astunparse==1.6.3
async-generator==1.10
attrs==21.1.0
autopep8==1.6.0
backcall==0.2.0
backports.zoneinfo==0.2.1
bcrypt==3.1.7
bleach==3.3.0
blinker==1.4
branca==0.5.0
Brlapi==0.7.0
cachetools==4.2.4
certifi==2019.11.28
cffi==1.14.5
chardet==3.0.4
chrome-gnome-shell==0.0.0
click==8.0.3
cloudpickle==2.2.0
colorama==0.4.3
command-not-found==0.3
commonmark==0.9.1
cryptography==2.8
cupshelpers==1.0
cycler==0.10.0
dbus-python==1.2.16
decorator==4.4.2
defer==1.0.6
defusedxml==0.7.1
display==1.0.0
distlib==0.3.1
distro==1.4.0
distro-info===0.23ubuntu1
Django==2.2.12
duplicity==0.8.12.0
EasyProcess==1.1
entrypoint2==1.1
entrypoints==0.3
et-xmlfile==1.1.0
facenet-pytorch==2.5.2
fasteners==0.14.1
filelock==3.0.12
Flask==2.0.2
flatbuffers==2.0
folium==0.12.1.post1
fuse-grader==0.1.9
future==0.18.2
gast==0.4.0
gitdb==4.0.9
GitPython==3.1.27
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
gpg===1.13.1-unknown
graphviz==0.20
grpcio==1.43.0
gunicorn==20.1.0
gym==0.26.2
gym-anytrading==1.3.2
gym-notices==0.0.8
h5py==3.6.0
htmlmin==0.1.12
httplib2==0.14.0
idna==2.8
ImageHash==4.3.1
importlib-metadata==4.12.0
ipykernel==5.5.4
ipyplot==1.1.1
ipython==7.22.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
itsdangerous==2.0.1
jedi==0.18.0
jeepney==0.8.0
Jinja2==3.0.3
joblib==1.1.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.12
jupyter-console==6.4.0
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
keras==2.7.0
Keras-Preprocessing==1.1.2
keyring==18.0.1
kiwisolver==1.3.1
kmodes==0.12.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lesscpy==0.14.0
libclang==12.0.0
lockfile==0.12.2
louis==3.12.0
lxml==4.6.3
macaroonbakery==1.3.1
Mako==1.1.0
Markdown==3.1.1
MarkupSafe==2.0.1
matplotlib==3.4.2
missingno==0.5.1
mistune==0.8.4
monotonic==1.5
mpl-finance==0.10.1
mplfinance==0.12.7a17
mss==7.0.1
multimethod==1.8
nbclient==0.5.4
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
netifaces==0.10.4
networkx==2.8.7
nltk==3.6.5
notebook==6.4.4
numpy==1.19.3
oauthlib==3.1.0
olefile==0.46
opencv-python==4.5.4.60
openpyxl==3.0.7
opt-einsum==3.3.0
packaging==21.3
pandas==1.2.4
pandas-datareader==0.10.0
pandas-profiling==3.3.0
pandocfilters==1.4.3
paramiko==2.6.0
parso==0.8.2
patsy==0.5.2
pbr==5.5.1
pexpect==4.6.0
phik==0.12.2
pickle-mixin==1.0.2
pickleshare==0.7.5
Pillow==7.0.0
pip-autoremove==0.10.0
playsound==1.2.2
ply==3.11
prometheus-client==0.10.1
prompt-toolkit==3.0.18
protobuf==3.19.1
psutil==5.9.1
ptyprocess==0.7.0
pyarrow==8.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.16.2
pycodestyle==2.8.0
pycparser==2.20
pycrypto==2.6.1
pycups==1.9.73
pydantic==1.9.2
pydeck==0.7.1
pygame==2.0.2
Pygments==2.9.0
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
Pympler==1.0.1
PyNaCl==1.3.0
pyparsing==2.4.6
PyQt5==5.14.1
pyRFC3339==1.1
pyrsistent==0.17.3
pyscreenshot==3.0
python-apt==2.0.1
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
python-libxdo==0.1.2a1
pytz==2019.3
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.4.1
pyxdg==0.26
PyYAML==5.3.1
pyzmq==22.0.3
qtconsole==5.1.0
QtPy==1.9.0
regex==2021.11.10
reportlab==3.5.34
requests==2.25.0
requests-oauthlib==1.3.0
requests-unixsocket==0.2.0
rich==12.5.1
rsa==4.8
scikit-learn==0.24.2
scipy==1.6.3
seaborn==0.11.1
SecretStorage==2.3.1
semver==2.13.0
Send2Trash==1.5.0
shortuuid==1.0.11
simplejson==3.16.0
sip==4.19.21
six==1.14.0
sklearn==0.0
smmap==5.0.0
sqlparse==0.2.4
statsmodels==0.13.2
streamlit==1.11.1
style==1.1.6
systemd-python==234
tangled-up-in-unicode==0.2.0
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.7.0
tensorflow-estimator==2.7.0
tensorflow-io-gcs-filesystem==0.23.1
termcolor==1.1.0
terminado==0.9.4
testpath==0.4.4
testresources==2.0.1
threadpoolctl==2.1.0
toml==0.10.2
toolz==0.12.0
torch==1.9.1+cpu
torchvision==0.10.1+cpu
tornado==6.1
tqdm==4.62.3
traitlets==5.0.5
typing_extensions==4.0.1
tzdata==2022.1
tzlocal==4.2
ubuntu-advantage-tools==27.12
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
usb-creator==0.3.7
validators==0.20.0
virtualenv==20.4.3
visions==0.7.5
wadllib==1.3.3
watchdog==2.1.9
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.2
whitenoise==5.2.0
widgetsnbextension==3.5.1
word2number==1.1
wrapt==1.13.3
xgboost==1.6.1
xkit==0.0.0
xlib==0.21
xlrd==2.0.1
zipp==3.8.1
